`get_rates()`: move to invoicing command to avoid circular import errors
